made changes for subdomaining; added root checking

+++ setupstack.sh
@@ -1,3 +1,8 @@
+if [[ $EUID -ne 0 ]]; then
+   echo "This script must be run as root" 
+   exit 1
+fi
+
 # retrieve resources from some remote server
 apt-get install -y git;
 rm -R ea-script;
